+++ cmd/carpenter.go
@@ -10,6 +10,7 @@ import (
 	"io"
 	"os"
 
+	"github.com/drgo/errors"
 	"github.com/drgo/fileutils"
 	"github.com/drgo/rosewood"
 )
@@ -19,90 +20,163 @@ var (
 	Version string
 	//Build holds the exe build number initialized in the Makefile
 	Build string
-	//Settings holds arguments/flags passed by the user
-	settings *rosewood.Settings
 )
 
-func init() {
-	settings = rosewood.DefaultSettings()
-	settings.ExecutableVersion = Version
-	settings.InputFileName = "<stdin>" //default to stdin
-
-	flag.StringVar(&settings.OutputFileName, "o", "", "")
-	flag.StringVar(&settings.OutputFileName, "output", "", "")
-
-	flag.StringVar(&settings.StyleSheetName, "css", "", "")
-
-	flag.BoolVar(&settings.Debug, "v", false, "")
-	flag.BoolVar(&settings.Debug, "verbose", false, "")
-
-	flag.StringVar(&settings.SectionSeparator, "s", settings.SectionSeparator, "")
-	flag.StringVar(&settings.SectionSeparator, "sep", settings.SectionSeparator, "")
-
-	flag.BoolVar(&settings.OverWriteOutputFile, "r", false, "")
-	flag.BoolVar(&settings.OverWriteOutputFile, "replace", false, "")
-
-	flag.BoolVar(&settings.CheckSyntaxOnly, "c", false, "")
-	flag.BoolVar(&settings.CheckSyntaxOnly, "check", false, "")
-}
-
 func main() {
-	flag.Usage = helpMessage
-	flag.Parse()
-	if settings.Debug {
-		fmt.Printf("current settings:\n")
-		fmt.Printf("%s\n", settings)
-	}
-	if flag.NArg() == 0 { //no arguments
-		if info, _ := os.Stdin.Stat(); info.Size() == 0 { //no input is being piped in
-			usage(1)
-		}
-	}
-	if err := Run(settings, flag.Args()); err != nil {
+	if err := RunApp(); err != nil {
 		crash(err)
 	}
-	os.Exit(0)
+}
+
+//RunApp has all program logic; easier to test than func main
+func RunApp() error {
+	if len(os.Args) == 1 { //only app name passed, exit
+		return fmt.Errorf(ErrWrongCommand)
+	}
+	settings := rosewood.DefaultSettings()
+	settings.ExecutableVersion = Version
+	settings.InputFileName = "<stdin>" //default to stdin
+	flgSets, _ := setupCommandFlag(settings)
+	flg, err := MustParseCommandLine(flgSets[0], flgSets[1:]...)
+	if err != nil {
+		return err
+	}
+	if settings.Debug {
+		fmt.Printf("current settings:\n%s\n", settings)
+	}
+	switch flg.Name() {
+	case "check":
+		settings.CheckSyntaxOnly = true
+		fallthrough
+	case "run":
+		if err := Run(settings, flg.Args()); err != nil {
+			return err
+		}
+	case "version":
+		fmt.Println(getVersion())
+	case "help":
+		helpMessage(flg.Args(), getVersion())
+	default:
+		helpMessage(nil, getVersion())
+		return fmt.Errorf(ErrWrongCommand)
+	}
+	return err
+}
+
+func setupCommandFlag(settings *rosewood.Settings) (flgSets []*flag.FlagSet, err error) {
+	globals := NewCommand("", []Flag{
+		{&settings.Debug, "verbose", "v", false},
+	})
+	cmdRun := NewCommand("run", []Flag{
+		{&settings.StyleSheetName, "style", "s", ""},
+		{&settings.OutputFileName, "output", "o", ""},
+		{&settings.OverWriteOutputFile, "replace", "r", false},
+		{&settings.SectionSeparator, "sep", "S", "+++"},
+	})
+	cmdCheck := NewCommand("check", []Flag{
+	//{&settings.TemplateFileArg, "template", "t", ""},
+	})
+	cmdHelp := NewCommand("help", []Flag{})
+	cmdVersion := NewCommand("version", []Flag{})
+	flgSets = append(flgSets, globals, cmdRun, cmdCheck, cmdHelp, cmdVersion)
+	for _, fs := range flgSets {
+		fs.Usage = func() {}    //disable internal usage function
+		fs.SetOutput(devNull{}) //suppress output from package flag
+	}
+	return flgSets, nil
 }
 
 //Run work-horse and main entry function
 func Run(settings *rosewood.Settings, args []string) error {
+	fmt.Println(len(args))
 	switch len(args) {
 	case 0: //input=stdin
 		if info, _ := os.Stdin.Stat(); info.Size() == 0 {
-			return fmt.Errorf("nothing is piped in")
+			return fmt.Errorf(ErrMissingInFile)
 		}
+		settings.InputFileName = "<stdin>"
 		if err := runSingle(settings); err != nil {
-			return fmt.Errorf("error running file [%s]: %s", "<stdin>", err)
+			return err //fmt.Errorf(ErrRunningFile, "<stdin>", err)
 		}
 	case 1: //input=single file
 		settings.InputFileName = args[0]
 		if err := runSingle(settings); err != nil {
-			return fmt.Errorf("error running file [%s]: %s", settings.InputFileName, err)
+			return err // fmt.Errorf(ErrRunningFile, settings.InputFileName, err)
 		}
 	default: //input= > 1 file
 		switch settings.OutputFileName {
 		case "": //one outputfile for each input file
-			var retErr error
+			errs := errors.NewErrorList()
 			for _, settings.InputFileName = range args {
 				if err := runSingle(settings); err != nil {
-					retErr = err
+					errs.Add(errors.ErrorsToError(err))
 				}
 			}
-			return fmt.Errorf("one or more errors running batch: %s", retErr)
+			return fmt.Errorf(ErrRunningBatch, errors.ErrorsToError(errs))
 		default: //all output goes into a single file
 			if err := runMulti(args, settings); err != nil {
-				return fmt.Errorf("one or more errors running batch: %s", err)
+				return fmt.Errorf(ErrRunningBatch, errors.ErrorsToError(err))
 			}
 		}
 	}
 	return nil
 }
 
-func runMulti(inFileNames []string, settings *rosewood.Settings) error {
+func runSingle(settings *rosewood.Settings) error {
 	var (
 		in          io.ReadCloser
 		out         io.WriteCloser
-		err, retErr error
+		err         error
+		minFileSize = settings.SectionsPerTable * len(settings.SectionSeparator)
+	)
+	annotateError := func(err error) error {
+		if err == nil {
+			return nil
+		}
+		return fmt.Errorf("error running file [%s]:\n%s", settings.InputFileName, err)
+	}
+	switch settings.InputFileName {
+	case "<stdin>":
+		in, _ = getInputFile("", minFileSize)
+		//output either stdout if outFileName=="" or outFileName
+		if !settings.CheckSyntaxOnly { //do not need an output for only checking syntax
+			if out, err = getOutputFile(settings.OutputFileName, settings.OverWriteOutputFile); err != nil {
+				return annotateError(err)
+			}
+			defer out.Close()
+		}
+	default: //single file
+		if in, err = getInputFile(settings.InputFileName, minFileSize); err != nil {
+			return annotateError(err)
+		}
+		defer in.Close()
+		//output either outFileName or a new file =inFileName + "ext" if outFileName==""
+		if !settings.CheckSyntaxOnly { //do not need an output
+			if settings.OutputFileName == "" {
+				settings.OutputFileName = fileutils.ReplaceFileExt(settings.InputFileName, "html")
+			}
+			if out, err = getOutputFile(settings.OutputFileName, settings.OverWriteOutputFile); err != nil {
+				return annotateError(err)
+			}
+			defer out.Close()
+		}
+	}
+	ri := rosewood.NewInterpreter(settings)
+	return annotateError(runFile(ri, in, out))
+}
+
+func runFile(ri *rosewood.Interpreter, in io.Reader, out io.Writer) error {
+	file, err := ri.Parse(bufio.NewReader(in), "")
+	if err != nil || ri.Settings().CheckSyntaxOnly {
+		return ri.ReportError(err)
+	}
+	return ri.ReportError(ri.RenderTables(out, file.Tables(), rosewood.NewHTMLRenderer()))
+}
+
+func runMulti(inFileNames []string, settings *rosewood.Settings) error {
+	var (
+		out         io.WriteCloser
+		err         error
 		minFileSize = settings.SectionsPerTable * len(settings.SectionSeparator)
 	)
 	//open output file if needed
@@ -113,84 +187,19 @@ func runMulti(inFileNames []string, settings *rosewood.Settings) error {
 		defer out.Close()
 	}
 	ri := rosewood.NewInterpreter(settings)
+	errs := errors.NewErrorList()
 	for _, f := range inFileNames {
-		in, err = getInputFile(f, minFileSize)
+		in, err := getInputFile(f, minFileSize)
 		if err != nil {
-			fmt.Printf("error opening file %s: %s\n", f, err)
-			retErr = err
+			errs.Add(err)
 			continue
 		}
 		defer in.Close()
-		err = runFile(ri, in, out)
-		report(settings, err)
-		if err != nil {
-			retErr = err
+		if err = runFile(ri, in, out); err != nil {
+			errs.Add(fmt.Errorf("error running file %s:\n%s", f, errors.ErrorsToError(err)))
 		}
 	}
-	return retErr
-}
-
-func runSingle(settings *rosewood.Settings) error {
-	var (
-		in          io.ReadCloser
-		out         io.WriteCloser
-		err         error
-		minFileSize = settings.SectionsPerTable * len(settings.SectionSeparator)
-	)
-	switch settings.InputFileName {
-	case "<stdin>":
-		in, _ = getInputFile("", minFileSize)
-		//output either stdout if outFileName=="" or outFileName
-		if !settings.CheckSyntaxOnly { //do not need an output
-			if out, err = getOutputFile(settings.OutputFileName, settings.OverWriteOutputFile); err != nil {
-				return err
-			}
-			defer out.Close()
-		}
-	default: //single file
-		if in, err = getInputFile(settings.InputFileName, minFileSize); err != nil {
-			return err
-		}
-		defer in.Close()
-		//output either outFileName or a new file =inFileName + "ext" if outFileName==""
-		if !settings.CheckSyntaxOnly { //do not need an output
-			if settings.OutputFileName == "" {
-				settings.OutputFileName = fileutils.ReplaceFileExt(settings.InputFileName, "html")
-			}
-			if out, err = getOutputFile(settings.OutputFileName, settings.OverWriteOutputFile); err != nil {
-				return err
-			}
-			defer out.Close()
-		}
-	}
-	ri := rosewood.NewInterpreter(settings)
-	err = runFile(ri, in, out)
-	return err
-}
-
-func runFile(ri *rosewood.Interpreter, in io.Reader, out io.Writer) error {
-	file, err := ri.Parse(bufio.NewReader(in), "")
-	if err != nil || ri.Settings().CheckSyntaxOnly {
-		return err
-	}
-	err = ri.RenderTables(out, file.Tables(), rosewood.NewHTMLRenderer())
-	report(ri.Settings(), err)
-	return err
-}
-
-func report(settings *rosewood.Settings, err error) {
-	if !settings.Debug {
-		return
-	}
-	out := settings.OutputFileName
-	if out == "" {
-		out = "<stdout>"
-	}
-	if err == nil {
-		fmt.Printf("File %s processed to %s\n", settings.InputFileName, out)
-	} else {
-		fmt.Printf("File %s failed\n", settings.InputFileName)
-	}
+	return errs
 }
 
 func getOutputFile(fileName string, overWrite bool) (*os.File, error) {
@@ -199,7 +208,7 @@ func getOutputFile(fileName string, overWrite bool) (*os.File, error) {
 	}
 	out, err := fileutils.CreateFile(fileName, overWrite)
 	if err != nil {
-		return nil, fmt.Errorf("error opening output file: %s", err)
+		return nil, fmt.Errorf(ErrOpenOutFile, fileName, err)
 	}
 	return out, nil
 }
@@ -220,30 +229,13 @@ func getInputFile(fileName string, minFileSize int) (*os.File, error) {
 	//rewind file stream
 	_, err = in.Seek(0, 0)
 	if err != nil {
-		return nil, fmt.Errorf("error opening input file %s: %s", fileName, err)
+		return nil, fmt.Errorf(ErrOpenInFile, fileName, err)
 	}
 	return in, nil
 }
 
-func helpMessage() {
-	fmt.Fprintf(os.Stderr, versionMessage, Version, Build)
-	fmt.Fprintln(os.Stderr, usageMessage)
-}
-
-func usage(exitCode int) {
-	helpMessage()
-	if exitCode > -1 {
-		os.Exit(exitCode)
-	}
-}
-
-const (
-	redColor = "\x1b[31m"
-)
-
-func crash(err error) {
-	fmt.Fprintf(os.Stderr, redColor+"%s\n", err)
-	os.Exit(1)
+func getVersion() string {
+	return fmt.Sprintf(versionMessage, Version, Build, rosewood.Version)
 }
 
 // func interactive() {