+++ changelog.md
@@ -1,3 +1,10 @@
 ## 0.3.1
 -	Crash due to invalid cell coordinates in a merge or style command
 -	Default section separator did not default to +++
+
+## 0.3.5
+- many changes...
+- new command line style eg carpenter run filename -options instead of carpenter filename -options run
+- changed most tests to fit the new style.
+- reading from stdin is disabled.
+- streamlined error messages and error reporting